+++ src/forms/Form.jsx
@@ -1,5 +1,6 @@
 import React, { useRef, useState } from 'react';
 import PropTypes from 'prop-types';
+
 import { FormContext } from './helpers/form-helpers';
 import { useForm } from './helpers/useForm';
 import { FormActions } from './FormActions';
@@ -22,7 +23,10 @@ export function Form({
   const [isSubmiting, setIsSubmiting] = useState(false);
 
   function resetForm() {
-    formRef.current.classList.remove('was-validated');
+    if (formRef.current && formRef.current.classList) {
+      formRef.current.classList.remove('was-validated');
+    }
+
     formState.reset();
   }