+++ src/forms/helpers/useFormControl.js
@@ -1,7 +1,7 @@
 import { useContext, useCallback } from 'react';
-import { isEmptyStringLike, isBoolean, isFunction } from 'js-var-type';
-import { FormContext } from './form-helpers';
-import { toDatetimeLocal, fromDatetimeLocal } from '../../utils/formatters';
+import { isFunction } from 'js-var-type';
+
+import { FormContext, getTargetValue, decode, encode } from './form-helpers';
 
 export function useFormControl(name, type) {
   const formState = useContext(FormContext);
@@ -50,73 +50,4 @@ export function useFormControl(name, type) {
     isValid: () => formState.getValidationMessage(name) === '',
     getFormSubmitedAttempted: () => formState.getSubmitedAttempted(),
   };
-}
-
-function getEmptyValue(type) {
-  switch (type) {
-    case 'boolean':
-      return false;
-
-    case 'array':
-      return [];
-
-    default:
-      return '';
-  }
-}
-
-function encode(value, type) {
-  if (isEmptyStringLike(value)) {
-    return getEmptyValue(type);
-  }
-
-  if (type === 'datetime-local') {
-    return toDatetimeLocal(value);
-  }
-
-  if (type === 'number' && isNaN(value)) {
-    return;
-  }
-
-  return value;
-}
-
-function decode(value, type) {
-  if (type === 'number') {
-    return parseFloat(value);
-  }
-
-  if (type === 'boolean') {
-    return isBoolean(value) ? value : value === 'true';
-  }
-
-  return value;
-}
-
-function getTargetValue(target) {
-  let value = target.type === 'checkbox' ? target.checked : target.value;
-
-  if (target.type === 'number') {
-    value = target.valueAsNumber;
-    if (isNaN(value)) {
-      value = undefined;
-    }
-  }
-
-  if (target.type === 'datetime-local') {
-    value = fromDatetimeLocal(target.value);
-  }
-
-  if (target.type === 'select-one') {
-    if (value && ['{', '['].includes(value[0])) {
-      try {
-        value = JSON.parse(value);
-      } catch (error) {
-        // eslint-disable-next-line no-console
-        console.error(error);
-      }
-    }
-  }
-
-  return value;
 }